Create symptom_trends_analysis.ipynb
This is will pull symptom data for the 4 weeks prior to recognition of an outbreak for measles and monkeypox and compares the same symptoms to regions where an outbreak didn't happen.  The goal is to determine if the symptoms being out of proportion to other described pathogens might suggest a new outbreak of an uncommon pathogen.